Mp3 bot (2)/requirements.txt
python-telegram-bot==20.3
yt-dlp==2023.10.07
ffmpeg-python==0.2.0